requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
aiohttp>=3.8.0
asyncio-mqtt>=0.16.0
websockets>=11.0.0
redis>=4.5.0
influxdb>=5.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
schedule>=1.2.0
python-telegram-bot>=20.0.0
click>=8.1.0
colorama>=0.4.6
tqdm>=4.65.0
pytz>=2023.3
python-dateutil>=2.8.2
requests>=2.31.0
urllib3>=2.0.0
certifi>=2023.5.7
charset-normalizer>=3.1.0
idna>=3.4
aiosignal>=1.3.1
attrs>=23.1.0
frozenlist>=1.3.3
multidict>=6.0.4
yarl>=1.9.2
psutil>=5.9.0
prometheus-client>=0.17.0
plotly>=5.15.0
dash>=2.11.0
dash-bootstrap-components>=1.4.0

# pandas-ta for technical analysis (replaces TA-Lib)
pandas-ta>=0.3.14b0

# Logging
loguru>=0.7.0

# Configuration
pyyaml>=6.0

# Docker support
docker>=7.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Optional ML dependencies (lightweight)
lightgbm>=4.0.0
xgboost>=1.7.0